drivers/accel: add new kconfig and update MAINTAINERS
Add a new Kconfig for the accel subsystem. The Kconfig currently
contains only the basic CONFIG_ACCEL option that will be used to
decide whether to compile the accel registration code as part of the
drm core functionality.

I have taken the liberty to appropriate the dri-devel mailing list and
the dri-devel IRC channel for the accel subsystem.

Signed-off-by: Oded Gabbay <ogabbay@kernel.org>